# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CAASPP & ELPAC Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [CAASPP, ELPAC]
        - name: subject
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentRequest"
      responses:
        "201":
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{assessmentId}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment by ID
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentWithQuestions"
    delete:
      operationId: deleteAssessment
      tags: [assessments]
      summary: Delete an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /assessments/{assessmentId}/questions:
    get:
      operationId: listAssessmentQuestions
      tags: [questions]
      summary: List questions for an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: addQuestionToAssessment
      tags: [questions]
      summary: Add a question to an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Question added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview
      parameters:
        - name: classId
          in: query
          schema:
            type: string
        - name: assessmentType
          in: query
          schema:
            type: string
            enum: [CAASPP, ELPAC]
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/student/{studentId}:
    get:
      operationId: getStudentAnalytics
      tags: [analytics]
      summary: Get analytics for a specific student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAnalytics"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        grade:
          type: string
          nullable: true
        classIds:
          type: array
          items:
            type: string
          nullable: true
        className:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, username, name, email, role, createdAt]
...
    Assessment:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [CAASPP, ELPAC]
        subject:
          type: string
        grade:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: integer
          description: Duration in minutes
        questionCount:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        status:
          type: string
          enum: [active, draft, archived]
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, subject, grade, duration, questionCount, difficulty, status, createdAt]
    AssessmentWithQuestions:
      allOf:
        - $ref: "#/components/schemas/Assessment"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
          required: [questions]
    CreateAssessmentRequest:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [CAASPP, ELPAC]
        subject:
          type: string
        grade:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
      required: [title, type, subject, grade, duration, difficulty]
    Question:
      type: object
      properties:
        id:
          type: string
        assessmentId:
          type: string
        text:
          type: string
        audioScript:
          type: string
          nullable: true
          description: Optional transcript spoken by the AudioPlayer instead of the question text
        type:
          type: string
          enum: [multiple_choice, short_answer, essay, listening, speaking]
        options:
          type: array
          nullable: true
          items:
            type: string
        correctAnswer:
          type: string
          nullable: true
        explanation:
          type: string
          nullable: true
        points:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        skill:
          type: string
          nullable: true
          description: ELA/Math/Science skill or ELPAC language skill
        orderIndex:
          type: integer
      required: [id, assessmentId, text, type, points, difficulty, orderIndex]
    CreateQuestionRequest:
      type: object
      properties:
        text:
          type: string
        audioScript:
          type: string
          nullable: true
        type:
          type: string
          enum: [multiple_choice, short_answer, essay, listening, speaking]
        options:
          type: array
          nullable: true
          items:
            type: string
        correctAnswer:
          type: string
          nullable: true
        explanation:
          type: string
          nullable: true
        points:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        skill:
          type: string
          nullable: true
        orderIndex:
          type: integer
      required: [text, type, points, difficulty, orderIndex]
    Result:
      type: object
      properties:
        id:
          type: string
        assessmentId:
          type: string
        assessmentTitle:
          type: string
        assessmentType:
          type: string
          enum: [CAASPP, ELPAC]
        studentId:
          type: string
        studentName:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        passed:
          type: boolean
        timeSpent:
          type: integer
          description: Time spent in seconds
        completedAt:
          type: string
          format: date-time
      required: [id, assessmentId, assessmentTitle, assessmentType, studentId, studentName, score, maxScore, percentage, passed, completedAt]
...
    AnalyticsOverview:
      type: object
      properties:
        totalStudents:
          type: integer
        totalAssessments:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number
        caasppStats:
          $ref: "#/components/schemas/AssessmentTypeStats"
        elpacStats:
          $ref: "#/components/schemas/AssessmentTypeStats"
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        scoreDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ScoreBucket"
        subjectPerformance:
          type: array
          items:
            $ref: "#/components/schemas/SubjectPerformance"
      required: [totalStudents, totalAssessments, averageScore, passRate, caasppStats, elpacStats, recentResults, scoreDistribution, subjectPerformance]
    AssessmentTypeStats:
      type: object
      properties:
        totalStudents:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectPerformance"
      required: [totalStudents, averageScore, passRate, subjectBreakdown]
    SubjectPerformance:
      type: object
      properties:
        subject:
          type: string
        averageScore:
          type: number
        passRate:
          type: number
        studentCount:
          type: integer
      required: [subject, averageScore, passRate, studentCount]
    ScoreBucket:
      type: object
      properties:
        range:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [range, count, percentage]
    StudentAnalytics:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/User"
        totalAssessments:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        progressOverTime:
          type: array
          items:
            $ref: "#/components/schemas/ProgressPoint"
        strengthAreas:
          type: array
          items:
            type: string
        improvementAreas:
          type: array
          items:
            type: string
      required: [student, totalAssessments, averageScore, passRate, recentResults, progressOverTime, strengthAreas, improvementAreas]
    ProgressPoint:
      type: object
      properties:
        date:
          type: string
          format: date-time
        score:
          type: number
        assessmentTitle:
          type: string
        assessmentType:
          type: string
      required: [date, score, assessmentTitle, assessmentType]
...